RestDemo classes moved to a new 'demo' package. Please consider moving them into a separate project and reconsider the demo security-wise.
