+++ pathfinder/src/main/java/bg/softuni/pathfinder/demo/RestDemoService.java
@@ -1,4 +1,4 @@
-package bg.softuni.pathfinder.service;
+package bg.softuni.pathfinder.demo;
 
 import bg.softuni.pathfinder.model.User;
 
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface RestDemoService {
 
-    List<User> getAll();
+    String getAll();
 }
