--- pathfinder/src/main/java/bg/softuni/pathfinder/service/impl/RestDemoServiceImpl.java
@@ -1,20 +1,0 @@
-package bg.softuni.pathfinder.service.impl;
-
-import bg.softuni.pathfinder.model.User;
-import bg.softuni.pathfinder.repository.UserRepository;
-import bg.softuni.pathfinder.service.RestDemoService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class RestDemoServiceImpl implements RestDemoService {
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public List<User> getAll() {
-        return userRepository.findAll();
-    }
-}
